pip-requirements.txt
BeautifulSoup==3.2.0
Django==1.3
PIL==1.1.7
South==0.7.3
akismet==0.2.0
django-blog-zinnia==0.8
django-mptt==0.4.2
django-tagging==0.3.1
django-xmlrpc==0.1.3
pyparsing==1.5.5
wsgiref==0.1.2
